ALL: Add setting foreignVatClasses on config form. Test with WC3
WC3: Since WC3, the shipping tax class can be "inherited" from the product items (which should be the preferred value for this setting). The plugin did not handle this value correctly.
WC3: Code to get the vat rate lookup meta data did not use customer location and would therefore be based on shop location.