WC|VM|HS: Use PluginConfig::VatClass_Null when no tax class is found when looking up the vat (class) meta data.
HS: HS 4+ has a setting on shipping methods "Automatic taxes"/"Automatische belastingen" that determines how the vat on shipping is calculated. The value "Proportion"/"Verhouding" uses the (weighed) average vat rate on the item lines as vat rate for the shipping. This component now supports this value. Note that it may result in multiple shipping lines if there are different vat rates on the item lines.