ALL: Error in config::getHostName(): using non-bracketed assignment operator in if condition. Minor consequences.
ALL: Gave the Config::Nature_... constants other values, so they can be treated as combinable bit values, including an upgrade function.
ALL: Updated code documentation to use the term "EU vat" instead of "Foreign EU vat".
ALL: Added support for the API v5.0 changes to the vat_info API call: the 'countryregion' is now also returned, allowing once more to decide if a country is EU or not. Older versions of this library will now erroneously decide that GB is EU.
ALL: Completor::getVatRatesByCountryAndInvoiceDate() now returns an array of vat infos instead of an array of vat rates.
ALL: Result::simplifyResponse() now changed to return an empty list when a list is expected but the main repsonse key was not found: This probably indicates an error but that should be found out e.g by checking the status or the messages. simplifyResponse() is aninternal method but influences the return value of Result::getResponse().